Add 'rm' function to filesystem modules and update package.json
The commit adds asynchronous and synchronous 'rm' (remove) functions in various filesystem modules within the project. The 'rm' function is created for both Promise and synchronous operations. In addition, the package.json file is updated to specify the use of "yarn@4.3.0" as the package manager. Minor changes are also made to other project configurations.

Signed-off-by: Crash <1903291+hex22a@users.noreply.github.com>